renamed paymentIntent folder to checkoutController
